Refacored and added parsers as a seperate section
Parsers are the set of objects that will parse all of the different sections of the page and get useful output from them
